objects/draw: move Object_DrawUnclippedItem to TRX
